refactor: move ci into cli
