+++ packages/cli/src/commands/ci/index.ts
@@ -1,8 +1,8 @@
 import {Command, flags} from '@heroku-cli/command'
 import * as Heroku from '@heroku-cli/schema'
 
-import {getPipeline} from '../../utils/pipelines'
-import {renderList} from '../../utils/test-run'
+import {getPipeline} from '@heroku-cli/plugin-ci/lib/utils/pipelines'
+import {renderList} from '@heroku-cli/plugin-ci/lib/utils/test-run'
 
 export default class CiIndex extends Command {
   static description = 'display the most recent CI runs for the given pipeline'
