+++ packages/cli/src/commands/ci/info.ts
@@ -1,8 +1,8 @@
 import {Command, flags} from '@heroku-cli/command'
 import * as Heroku from '@heroku-cli/schema'
 
-import {getPipeline} from '../../utils/pipelines'
-import {displayTestRunInfo} from '../../utils/test-run'
+import {getPipeline} from '@heroku-cli/plugin-ci/lib/utils/pipelines'
+import {displayTestRunInfo} from '@heroku-cli/plugin-ci/lib/utils/test-run'
 
 export default class CiInfo extends Command {
   static description = 'show the status of a specific test run'
