+++ packages/cli/test/unit/commands/ci/info.unit.test.ts
